+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import prevent_app_screen
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  PreventAppScreenPlugin.register(with: registry.registrar(forPlugin: "PreventAppScreenPlugin"))
+}