+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -1,0 +1,14 @@
+//
+//  Generated file. Do not edit.
+//
+
+// clang-format off
+
+#include "generated_plugin_registrant.h"
+
+#include <prevent_app_screen/prevent_app_screen_plugin.h>
+
+void RegisterPlugins(flutter::PluginRegistry* registry) {
+  PreventAppScreenPluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("PreventAppScreenPlugin"));
+}