+++ linux/prevent_app_screen_plugin_private.h
@@ -1,0 +1,9 @@
+#include <flutter_linux/flutter_linux.h>
+
+#include "include/prevent_app_screen/prevent_app_screen_plugin.h"
+
+// This file exposes some plugin internals for unit testing. See
+// https://github.com/flutter/flutter/issues/88724 for current limitations
+// in the unit-testable API.
+
+// Internal method handlers would go here.